Move ColumnId to interfaces
