+++ src/table/constants.ts
@@ -1,6 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export type ColumnId = string | symbol;
 export const SELECTION_CELL_WIDTH = 54;
 export const selectionColumnId = Symbol('selection-column-id');
