+++ src/table/header-cell/index.tsx
@@ -13,7 +13,7 @@ import { InteractiveComponent } from '../thead';
 import { StickyColumnsModel, useStickyCellStyles } from '../use-sticky-columns';
 import { getStickyClassNames } from '../utils';
 import { useInternalI18n } from '../../internal/i18n/context';
-import { ColumnId } from '../constants';
+import { ColumnId } from '../interfaces';
 
 interface TableHeaderCellProps<ItemType> {
   className?: string;
