+++ src/table/interfaces.tsx
@@ -457,3 +457,5 @@ export namespace TableProps {
     visible: boolean;
   }
 }
+
+export type ColumnId = string | symbol;
